Move tab title field from settings panel to frames panel
